find a bug
When I use the Multiplexing Channel Selector in apache-flume-ng-1.9.0, I found that the topicHeader parameter specified in KafkaSource does not work properly for mapping. Refer to the source code of kafkasource and found that in line 266, the key of the map that should put into the header is a dynamic parameter,  but in my humbly opinion , this should be a static string ("topicHeader"). when I modify the source code, the selector works normally !